refactor: 컨트롤러/예외 핸들러 내부 record 제거 및 코드 정리
- 내부 AnalysisIdResponse 제거
- 중복 로그 제거
- 내부 ErrorResponse / ParsingErrorResponse 제거
- 불필요한 import 정리
