+++ src/main/java/com/electricip/loganalyzer/api/GlobalExceptionHandler.java
@@ -10,7 +10,6 @@ import com.electricip.loganalyzer.domain.exception.LogParsingException;
 import com.electricip.loganalyzer.domain.exception.TooManyParsingErrorsException;
 import com.electricip.loganalyzer.infrastructure.client.IpInfoException;
 import com.electricip.loganalyzer.infrastructure.client.RateLimitExceededException;
-import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
@@ -21,9 +20,6 @@ import org.springframework.web.bind.MissingServletRequestParameterException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.multipart.MaxUploadSizeExceededException;
-
-import java.time.LocalDateTime;
-import java.util.List;
 
 /**
  * 전역 예외 처리기
@@ -334,97 +330,4 @@ public class GlobalExceptionHandler {
                 ));
     }
 
-    /**
-     * 에러 응답 (Record)
-     */
-    @Schema(description = "에러 응답")
-    public record ErrorResponse(
-            @Schema(description = "에러 발생 시각", example = "2026-02-09T14:30:00")
-            LocalDateTime timestamp,
-            @Schema(description = "HTTP 상태 코드", example = "400")
-            int status,
-            @Schema(description = "에러 코드", example = "INVALID_FILE")
-            String errorCode,
-            @Schema(description = "에러 메시지", example = "파일이 비어있습니다")
-            String message,
-            @Schema(description = "요청 경로", example = "/api/analysis")
-            String path
-    ) {
-        /**
-         * Compact Constructor: 필수 필드 검증 + message null 치환
-         */
-        public ErrorResponse {
-            java.util.Objects.requireNonNull(timestamp, "timestamp는 null일 수 없습니다");
-            java.util.Objects.requireNonNull(errorCode, "errorCode는 null일 수 없습니다");
-            java.util.Objects.requireNonNull(path, "path는 null일 수 없습니다");
-            message = (message != null) ? message : "알 수 없는 오류";
-        }
-
-        public static ErrorResponse of(HttpStatus httpStatus, String errorCode,
-                                       String message, String path) {
-            return new ErrorResponse(
-                    LocalDateTime.now(),
-                    httpStatus.value(),
-                    errorCode,
-                    message,
-                    path
-            );
-        }
-    }
-
-    /**
-     * 파싱 에러 과다 응답 — ErrorResponse 필드 + 파싱 통계/샘플
-     */
-    @Schema(description = "파싱 에러 과다 응답 (422)")
-    public record ParsingErrorResponse(
-            @Schema(description = "에러 발생 시각", example = "2026-02-09T14:30:00")
-            LocalDateTime timestamp,
-            @Schema(description = "HTTP 상태 코드", example = "422")
-            int status,
-            @Schema(description = "에러 코드", example = "TOO_MANY_PARSING_ERRORS")
-            String errorCode,
-            @Schema(description = "에러 메시지", example = "유효한 로그가 없습니다 (전체 100줄 중 100줄 에러)")
-            String message,
-            @Schema(description = "요청 경로", example = "/api/analysis")
-            String path,
-            @Schema(description = "전체 라인 수", example = "100")
-            long totalLines,
-            @Schema(description = "에러 라인 수", example = "100")
-            long errorCount,
-            @Schema(description = "에러 샘플 (최대 10건)")
-            List<ParseErrorSample> errorSamples
-    ) {
-        public ParsingErrorResponse {
-            java.util.Objects.requireNonNull(timestamp, "timestamp는 null일 수 없습니다");
-            java.util.Objects.requireNonNull(errorCode, "errorCode는 null일 수 없습니다");
-            java.util.Objects.requireNonNull(path, "path는 null일 수 없습니다");
-            message = (message != null) ? message : "알 수 없는 오류";
-            errorSamples = (errorSamples != null) ? List.copyOf(errorSamples) : List.of();
-        }
-
-        public static ParsingErrorResponse of(String errorCode, String message, String path,
-                                               long totalLines, long errorCount,
-                                               List<ParseErrorSample> errorSamples) {
-            return new ParsingErrorResponse(
-                    LocalDateTime.now(),
-                    HttpStatus.UNPROCESSABLE_ENTITY.value(),
-                    errorCode,
-                    message,
-                    path,
-                    totalLines,
-                    errorCount,
-                    errorSamples
-            );
-        }
-
-        @Schema(description = "파싱 에러 샘플")
-        public record ParseErrorSample(
-                @Schema(description = "에러 발생 라인 번호", example = "42")
-                long lineNumber,
-                @Schema(description = "에러 메시지", example = "잘못된 날짜 형식")
-                String errorMessage,
-                @Schema(description = "에러 유형 (PARSING, VALIDATION, FORMAT)", example = "PARSING")
-                String errorType
-        ) {}
-    }
 }
