+++ src/test/java/com/electricip/loganalyzer/api/ErrorResponseTest.java
@@ -18,7 +18,7 @@ class ErrorResponseTest {
         @Test
         @DisplayName("모든 필드가 정상이면 그대로 생성된다")
         void shouldCreateWithValidFields() {
-            var response = new GlobalExceptionHandler.ErrorResponse(
+            var response = new ErrorResponse(
                     LocalDateTime.of(2025, 1, 1, 0, 0),
                     400, "INVALID_REQUEST", "잘못된 요청", "/api/test");
 
@@ -31,7 +31,7 @@ class ErrorResponseTest {
         @Test
         @DisplayName("message가 null이면 '알 수 없는 오류'로 치환된다")
         void shouldReplaceNullMessageWithDefault() {
-            var response = new GlobalExceptionHandler.ErrorResponse(
+            var response = new ErrorResponse(
                     LocalDateTime.now(), 400, "CODE", null, "/api/test");
 
             assertEquals("알 수 없는 오류", response.message());
@@ -41,7 +41,7 @@ class ErrorResponseTest {
         @DisplayName("timestamp가 null이면 NullPointerException 발생")
         void shouldThrowWhenTimestampIsNull() {
             assertThrows(NullPointerException.class, () ->
-                    new GlobalExceptionHandler.ErrorResponse(
+                    new ErrorResponse(
                             null, 400, "CODE", "msg", "/api/test"));
         }
 
@@ -49,7 +49,7 @@ class ErrorResponseTest {
         @DisplayName("errorCode가 null이면 NullPointerException 발생")
         void shouldThrowWhenErrorCodeIsNull() {
             assertThrows(NullPointerException.class, () ->
-                    new GlobalExceptionHandler.ErrorResponse(
+                    new ErrorResponse(
                             LocalDateTime.now(), 400, null, "msg", "/api/test"));
         }
 
@@ -57,7 +57,7 @@ class ErrorResponseTest {
         @DisplayName("path가 null이면 NullPointerException 발생")
         void shouldThrowWhenPathIsNull() {
             assertThrows(NullPointerException.class, () ->
-                    new GlobalExceptionHandler.ErrorResponse(
+                    new ErrorResponse(
                             LocalDateTime.now(), 400, "CODE", "msg", null));
         }
     }
@@ -69,7 +69,7 @@ class ErrorResponseTest {
         @Test
         @DisplayName("정상 생성")
         void shouldCreateViaOf() {
-            var response = GlobalExceptionHandler.ErrorResponse.of(
+            var response = ErrorResponse.of(
                     HttpStatus.BAD_REQUEST, "INVALID_REQUEST", "잘못된 요청", "/api/test");
 
             assertEquals(400, response.status());
@@ -80,7 +80,7 @@ class ErrorResponseTest {
         @Test
         @DisplayName("of()에서 message가 null이면 기본 메시지로 치환된다")
         void shouldReplaceNullMessageInOf() {
-            var response = GlobalExceptionHandler.ErrorResponse.of(
+            var response = ErrorResponse.of(
                     HttpStatus.BAD_REQUEST, "CODE", null, "/api/test");
 
             assertEquals("알 수 없는 오류", response.message());
